# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample send request use API.",
        "title": "Application EventHub",
        "contact": {},
        "version": "1.0"
    },
    "host": "172.17.0.1:2027",
    "basePath": "/",
    "paths": {
...
        "/user/community/joined": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "for get list of events joined by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Joined Communities",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/events/joined": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "for get list of events joined by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Joined Events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/events/saved": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "for get list of events joined by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Saved Events",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/profiles": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "for get detail information profiles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get Profiles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/set/password": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Change password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Change Password",
                "parameters": [
                    {
                        "description": "Body to Change Password",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ReqNewPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        },
        "/user/set/profiles": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update user profile information and upload a profile picture (max 2MB, formats: jpg/jpeg/png)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update Profile",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Alma'ruf Hidayat",
                        "description": "User Full Name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "Masamba, Luwu Utara, Sulawesi Selatan",
                        "description": "User Address",
                        "name": "address",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "Fullstack Developer",
                        "description": "User Job Title",
                        "name": "job",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "PT. Best Life Ummah",
                        "description": "User Office",
                        "name": "office",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "Senior Fullstack Developer",
                        "description": "Description",
                        "name": "description",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Profile Picture File",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ResSetProfile"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "failed"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "dto.ReqNewPassword": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string",
                    "minLength": 8
                },
                "old_password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "dto.ResSetProfile": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "job": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "office": {
                    "type": "string"
                }
            }
        },
        "dto.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer Token used as identity for accessing backend",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}